#609 講師のチャプター関連APIで、講座作成した講師のみしかチャプターのステータス更新・削除できないように制限する(ごめ)

+++ app/Http/Controllers/Api/Instructor/ChapterController.php
@@ -107,9 +107,25 @@ class ChapterController extends Controller
      */
     public function updateStatus(ChapterPatchStatusRequest $request)
     {
-        Chapter::findOrFail($request->chapter_id)
-            ->update([
-                'status' => $request->status
+        $chapter = Chapter::with('course')->findOrFail($request->chapter_id);
+
+        if (Auth::guard('instructor')->user()->id !== $chapter->course->instructor_id) {
+            return response()->json([
+                'result' => false,
+                "message" => 'invalid instructor_id.'
+            ], 403);
+        }
+
+        if ((int) $request->course_id !== $chapter->course->id) {
+            // 指定した講座IDがチャプターの講座IDと一致しない場合は更新を許可しない
+            return response()->json([
+                'result'  => false,
+                'message' => 'Invalid course_id.',
+            ], 403);
+        }
+
+        $chapter->update([
+            'status' => $request->status
         ]);
 
         return response()->json([
@@ -125,8 +141,25 @@ class ChapterController extends Controller
      */
     public function delete(ChapterDeleteRequest $request)
     {
-        $chapter = Chapter::findOrFail($request->chapter_id);
+        $chapter = Chapter::with('course')->findOrFail($request->chapter_id);
+
+        if (Auth::guard('instructor')->user()->id !== $chapter->course->instructor_id) {
+            return response()->json([
+                'result' => false,
+                "message" => 'invalid instructor_id.'
+            ], 403);
+        }
+
+        if ((int) $request->course_id !== $chapter->course->id) {
+            // 指定した講座IDがチャプターの講座IDと一致しない場合は更新を許可しない
+            return response()->json([
+                'result'  => false,
+                'message' => 'Invalid course_id.',
+            ], 403);
+        }
+
         $chapter->delete();
+
         return response()->json([
             "result" => true
         ]);
